Move actions to Http namespace
