+++ CargoBackend/src/Http/Action/GetCargo.php
@@ -6,31 +6,29 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  * 
- * Date: 07.12.2015 - 8:39 PM
+ * Date: 07.12.2015 - 2:15 PM
  */
-namespace Codeliner\CargoBackend\Application\Action;
+namespace Codeliner\CargoBackend\Http\Action;
+
 use Codeliner\CargoBackend\Application\Booking\BookingService;
-use Codeliner\CargoBackend\Application\Booking\Dto\CargoRoutingDto;
+use Codeliner\CargoBackend\Application\Exception\CargoNotFoundException;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Zend\Diactoros\Response\EmptyResponse;
 use Zend\Diactoros\Response\JsonResponse;
 
 /**
- * Class GetCargos
+ * Class GetCargo
  *
  * @package Codeliner\CargoBackend\Application\Action
  */
-final class GetCargos
+final class GetCargo
 {
     /**
      * @var BookingService
      */
     private $bookingService;
 
-    /**
-     * GetCargos constructor.
-     * @param BookingService $bookingService
-     */
     public function __construct(BookingService $bookingService)
     {
         $this->bookingService = $bookingService;
@@ -38,8 +36,16 @@ final class GetCargos
 
     public function __invoke(ServerRequestInterface $request, ResponseInterface $response, callable $next)
     {
-        return new JsonResponse(['cargos' => array_map(function(CargoRoutingDto $cargoRoutingDto) {
-            return $cargoRoutingDto->getArrayCopy();
-        }, $this->bookingService->listAllCargos() )]);
+        if (null === $trackingId = $request->getAttribute('trackingId')) {
+            return new EmptyResponse(404);
+        }
+
+        try {
+            $cargoRoutingDto = $this->bookingService->loadCargoForRouting($trackingId);
+
+            return new JsonResponse($cargoRoutingDto->getArrayCopy());
+        } catch (CargoNotFoundException $e) {
+            return new EmptyResponse(404);
+        }
     }
 }
