+++ CargoBackend/src/Http/Action/UpdateCargo.php
@@ -8,7 +8,7 @@
  * 
  * Date: 07.12.2015 - 6:30 PM
  */
-namespace Codeliner\CargoBackend\Application\Action;
+namespace Codeliner\CargoBackend\Http\Action;
 
 use Codeliner\CargoBackend\Application\Booking\BookingService;
 use Codeliner\CargoBackend\Application\Booking\Dto\LegDto;
